requirements.txt
pyserial
SQLAlchemy